% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5_17_Package.R
\docType{package}
\name{swelection}
\alias{swelection}
\alias{swelection-package}
\title{Sweden election 2014}
\description{
This package is used to get the 2014 election results in Sweden for the nine biggest parties.The package contains two functions which are "votes_spec" and "votes_sum".
}
\author{
Ahmet Akdeve \email{ahmak554@student.liu.se}

Zhixuan Duan \email{darinstu999@gmail.com}
}
