% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votes_sum.R
\name{votes_sum}
\alias{votes_sum}
\title{Sweden election 2014}
\usage{
votes_sum(the_code)
}
\arguments{
\item{the_code}{A The municipality code}
}
\value{
A data frame with 20 columns containing in percentages for each party and 290 rows for each municipality in Sweden
}
\description{
The function returns the result per municipality. The function can take 3 values. "00K" gives the result in the municipal elections, "00L" gives the result in the county elections per municipality and "00R" gives the result in the national elections per municipality.
}
